new version
Former-commit-id: 1776ba8b5473e9666c78bfe540822eab0e9c1312 [formerly b2cf64577aeba99ea42100883df7479ea5c87c5f]
Former-commit-id: 8537a70c484982e6a6e4a2d9194895049e5b6302

+++ transactions/transactions.js
@@ -38,8 +38,8 @@ transaction.prototype.getBytes = function () {
         bb.writeByte(recepient[i] || 0);
     }
 
-    bb.writeInt(this.amount);
-    bb.writeInt(this.fee);
+    bb.writeFloat(this.amount);
+    bb.writeFloat(this.fee);
 
     if (this.referencedTransaction) {
         var referncedTransactionBuffer = bignum(this.referencedTransaction).toBuffer();
